refactor(managed): extract shared useIdentityFocus hook
The /managed hub's '?focus= by managed identity' filter logic (URL-synced
selection, focusedDid mapping, single-group via-suppression, segmented-vs-
dropdown choice) moves into a reusable hook so the notifications page can
share it verbatim. Pure refactor — the managed hub behaves identically
(browser-reverified).
